Research/curbsidepickup/src/alpr/lva_ai_solution/requirements.txt
pillow<7.0.0
click==6.7
configparser==3.5.0
Flask==0.12.2
gunicorn==19.6.0
json-logging-py==0.2
MarkupSafe==1.0
olefile==0.44
requests==2.18.4
six==1.13.0
opencv-python==4.1.1.26
imgaug==0.3.0
torch==1.2
cython==0.29.14
addict==2.2.1
azure-storage-blob==12.3.0
python-dotenv==0.13.0